added GetLocation to ConsequenceReceptor and changed Location coordinate types to double to match NSI #27

+++ Consequences/Geography/Location.cs
@@ -1,6 +1,6 @@
 ﻿namespace USACE.HEC.Geography;
 public struct Location
 {
-  public float X; 
-  public float Y;
+  public double X; 
+  public double Y;
 }
